Switch base image to alpine, move boot into a bin script
We've been running into some very strange issues with cron not
triggering properly under the Ubuntu base image. I had an image in a
reliable state where it would come and never execute anything until you
ran `crontab -e` or `-l` to "touch" the cron system in some way.
Investigating changing how we loaded the crontab was triggered by a
customer issue with unreliable builder-gc. In my testing the alpine
image has worked exactly as expected and not flaky at all, and it also
cuts down the image size by a couple hundred MB, so I think this is
worthile, although it does mean we need to lean on less-readable flags
for `xargs` and `tail`.